removed Tkinter ui
