--- src/speech_mcp/ui/pyqt/__init__.py
@@ -1,7 +1,0 @@
-"""
-PyQt implementation of the Speech UI.
-"""
-
-from .pyqt_ui import PyQtSpeechUI, AudioProcessor, AudioVisualizer, run_ui
-
-__all__ = ["PyQtSpeechUI", "AudioProcessor", "AudioVisualizer", "run_ui"]
